pass in an initialiser function.
in app engine slice items are appended on to the end of existing slices rather than clearing the existing elements first and then adding the retrieved slices elements into it. so we need to us the versionFactory to create empty structs with empty slices for entity retrieval, and use the entityInitialiser to initialise "naked" entities just in the create calls.